feat(openapi): find swagger spec in sub directories (PS-216)
Use case: we have a build where the openapi spec is produced in a
service's submodule build directory but we want to run the top level
build to also run the other submodules' tests.

This will find the spec in any subdirectory matching the target path
so we are assuming there is only one spec per build generated, which
I think is fair.