feat(kotlin): Service profile generation for kotlin (#154)
Add a new workflow that can be called from the deploy workflow and
revives service profile generation. In order not to have to connect to
teleport or headscale from github actions, we are taking the gitops
route to push the generated service profile to kube-manifests.

In order to provide some flexibility, it's possible to change the
service name after the file is created in case the deployed service has
a different name than the standard {identifier}-{stage} pattern.